% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecm.R
\name{AdaEN}
\alias{AdaEN}
\title{Perform Adaptive Elastic Net with optimal adaptive weights}
\usage{
AdaEN(
  x,
  y,
  weights = NULL,
  adaptive = TRUE,
  adaptive.gamma = 1,
  lower.limits = -Inf,
  upper.limits = Inf,
  alphas = c(0, 0.01, 0.05, seq(0.1, 1, 0.1)),
  lambda = NULL,
  nlambda = 51,
  parallel = FALSE,
  cluster = NULL,
  plot.cv = FALSE,
  time.series = c("no", "sliding", "expanding"),
  ts.out.sample = 10,
  se.mult = NULL,
  best.choice = c("min", "se"),
  full.relax = FALSE,
  ...
)
}
\arguments{
\item{x}{Passed to glmnet: a matrix of predictors without the constant (all finite values).}

\item{y}{Passed to glmnet: a numeric vector of dependent variable values (all finite).}

\item{weights}{Passed to glmnet: observation weights}

\item{adaptive}{If TRUE, use adaptive regressor weights}

\item{adaptive.gamma}{Adaptive weight power in 1/|beta_initial|^gamma; a positive number, usually 1, 2 or 0.5.}

\item{lower.limits}{Passed to glmnet: box constraints from below.}

\item{upper.limits}{Passed to glmnet: box constraints from above.}

\item{alphas}{Numeric vector of the sequence of elastic net mixing parameters (1 = LASSO, 0 = ridge).}

\item{lambda}{Optional: a pre-defined sequence of elastic net penalty lambdas if one alpha is chosen.}

\item{nlambda}{Passed to glmnet: the number of lambdas to choose over}

\item{parallel}{Logical: parallelise computations?}

\item{cluster}{A cluster for parallel computations to be passed to parLapplyLB.}

\item{plot.cv}{Logical: if TRUE, produce a contour plot of the cross-validation stage.}

\item{time.series}{Character indicating whether the CV should be done via LOO (default), sliding window, or expanding window}

\item{ts.out.sample}{If 'time.series' is 'expanding' or 'sliding', the number of points at the end to use for testing}

\item{se.mult}{In times series cross-validation, the margin for increasing the penalty within this number of forecast MSE SDs.}

\item{best.choice}{Which model is considered the best: the one that minimises MSE or a more parsimonious one within the 'se.mult' limit
See Chen & Yang (2021) The One Standard Error Rule for Model Selection: Does It Work? for a discussion (in short, "min" should be better).}

\item{full.relax}{Cross-validate with respect to the MSE from the relaxed fit?}

\item{...}{Passed to contour that is used for plotting.}
}
\value{
A list of optimal adaptive weights, estimated Elastic Net, cross-validated cv.glmnet, optimal coefficients, optimal alpha, and optimal lambda.
}
\description{
Perform Adaptive Elastic Net with optimal adaptive weights
}
\examples{
set.seed(1)
x1 <- arima.sim(list(ar = c(0.4, 0.3)), n = 100)
x2 <- arima.sim(list(ar = c(0.4, 0.3)), n = 100)
x3 <- arima.sim(list(ar = c(0.4, 0.3)), n = 100)
y <- 1 + x1 + x2 + rnorm(100)
x <- cbind(x1, x2, x3)
a <- AdaEN(x = x, y = y, plot.cv = TRUE, time.series = "expanding",
           ts.out.sample = 15, alphas = 0:5/5) # Takes 15 seconds to run
}
