% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calendars}
\alias{calendars}
\alias{Belgium.M}
\alias{Belgium.Q}
\alias{France.M}
\alias{France.Q}
\alias{Germany.M}
\alias{Germany.Q}
\alias{Luxembourg.M}
\alias{Luxembourg.Q}
\title{National calendars for regression trading-day pre-adjustment}
\format{
An object of class \code{mts} (inherits from \code{ts}, \code{matrix}) with 1680 rows and 8 columns.

An object of class \code{mts} (inherits from \code{ts}, \code{matrix}) with 560 rows and 8 columns.

An object of class \code{mts} (inherits from \code{ts}, \code{matrix}) with 1680 rows and 8 columns.

An object of class \code{mts} (inherits from \code{ts}, \code{matrix}) with 560 rows and 8 columns.

An object of class \code{mts} (inherits from \code{ts}, \code{matrix}) with 1680 rows and 8 columns.

An object of class \code{mts} (inherits from \code{ts}, \code{matrix}) with 560 rows and 8 columns.

An object of class \code{mts} (inherits from \code{ts}, \code{matrix}) with 1680 rows and 8 columns.

An object of class \code{mts} (inherits from \code{ts}, \code{matrix}) with 560 rows and 8 columns.
}
\usage{
Belgium.M

Belgium.Q

France.M

France.Q

Germany.M

Germany.Q

Luxembourg.M

Luxembourg.Q
}
\description{
These datasets contain the calendars for some European countries that were
obtained from JDemetra+ 2.2.3 by creating national holidays.
They are used in seasonal adjustment procedures (diagnoseSeasonality())
when the user requests a national calendar
(e.g. diagnoseSeasonality(x, calendar = "Luxembourg"))
}
\details{
The first part of the name is the country, the second denotes frequency (monthly or quarterly).

The source JD+ file is data-raw/All-National-Calendars-JDemetra.cfgx. It
requires a working
}
\keyword{datasets}
