% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributions.R
\name{compareContributions}
\alias{compareContributions}
\title{Compare two sets of contributions from multple models}
\usage{
compareContributions(old, new)
}
\arguments{
\item{old}{A list returned by computeContribManyLM}

\item{new}{A list returned by computeContribManyLM}
}
\value{
A list with the differences of contributions
}
\description{
Compare two sets of contributions from multple models
}
\examples{
set.seed(1)
x1 <- rnorm(100) + 1:100/50 + 1
x2 <- rnorm(100) - 1:100/50 - 1
x3 <- rnorm(100)
y <- 1 + x1 + x2 + 0.5*x3 + rnorm(100)
d <- data.frame(y, x1, x2, x3)
mod1 <- lm(y ~ x1 + x2, data = d)
mod2 <- lm(y ~ x1 + x3, data = d)
mod3 <- lm(y ~ x1 + x2 + x3, data = d)
d2 <- d[71:100, ]
ctb1 <- computeContribManyLM(list(mod1, mod2), newdata = d2)
ctb2 <- computeContribManyLM(list(mod1, mod2, mod3), newdata = d2)
ctb.diff <- compareContributions(ctb1, ctb2)
plotContribLM(ctb.diff, main = "Contribution discrepancies")
}
