% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributions.R
\name{computeContribLM}
\alias{computeContribLM}
\title{Compute contributions for a linear model}
\usage{
computeContribLM(
  model = NULL,
  coef = NULL,
  newdata = NULL,
  dep.var.name = NULL,
  resid.name = "resid",
  groups = NULL,
  zero.residual = FALSE,
  warn.on.na.depvar = FALSE
)
}
\arguments{
\item{model}{An lm object for which to compute the contributions}

\item{coef}{If there is no model, a named vector of coefficients obtained from any source. Requires non-empty 'newdata' with matching column names}

\item{newdata}{A new data frame for which to compute the contributions; by default, the training data set is used}

\item{dep.var.name}{If model is NULL and coef is not NULL, use this variable from newdata as the observed values}

\item{resid.name}{A character for residual label (cosmetic)}

\item{groups}{A list of character vectors determining which groups of regressors to add up}

\item{zero.residual}{Compute the contributions with or without the residual?}

\item{warn.on.na.depvar}{If TRUE and the RHS variable has NAs, wars about the zero-residual assumption.}
}
\value{
A list of two matrices: contributions to the levels and differences.
}
\description{
Compute contributions for a linear model
}
\examples{
set.seed(1)
x1 <- as.numeric(arima.sim(list(ar = c(0.4, 0.3)), n = 100))
x2 <- as.numeric(arima.sim(list(ar = c(0.4, 0.3)), n = 100))
y <- 1 + x1 + x2 + as.numeric(arima.sim(list(ma = 0.7), n = 100))
d <- data.frame(y, x1, x2)
mod <- lm(y ~ x1 + x2, data = d)
ctb <- computeContribLM(mod)
round(tail(ctb$level, 5), 3) # Contributions to Y according to the model
true.coef <- c(`(Intercept)` = 1, x1 = 1, x2 = 1)
ctb.true <- computeContribLM(coef = true.coef, newdata = d, dep.var.name = "y")
round(tail(ctb.true$level, 5), 3) # Real contributions to Y
plotContribLM(ctb.true, mar = c(0.2, 2, 2, 0.2), main = "Decomposition",
              left.bar.skip = 10, mtext.line = -1)
}
