% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{computeSpectrum}
\alias{computeSpectrum}
\title{Compute the power spectrum of a series}
\usage{
computeSpectrum(
  x,
  apply.diff = TRUE,
  pad = 2,
  return.half = TRUE,
  plot = TRUE,
  compact = FALSE,
  lwd = 1.5,
  put.legend = TRUE
)
}
\arguments{
\item{x}{A univariate ts object or a numeric vector}

\item{apply.diff}{If TRUE, takes the differences (standard in JDemetra+)}

\item{pad}{Pad the series with 0s up to this many orders of magnitude above the length of x. 0 or negative = no padding.}

\item{return.half}{If TRUE, return only the first half of the FFT (due to the symmetrical nature of the FFT, the second half mirrors the first one)}

\item{plot}{If TRUE, produces a plot}

\item{compact}{If TRUE, suppres the y axis and axis labels}

\item{lwd}{Spectrogram line width}

\item{put.legend}{If TRUE, adds a legend (that depends on the \code{compact} argument)

A useful tool to visualise seasonal spikes. The presence of spikes indicates that there are periodic phenomena in the data.}
}
\value{
A list of the frequency grid and corresponding frequency intensity.
}
\description{
Compute the power spectrum of a series
}
\examples{
computeSpectrum(datasets::AirPassengers) # Strong seasonal components
# Adjust the series by LOESS and remove the seasonal fluctuations
ap.sa <- stl(datasets::AirPassengers, 9)
computeSpectrum(ap.sa$time.series[, "remainder"]) # Strong trading-day effect!
}
