% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa.R
\name{dateAsEV}
\alias{dateAsEV}
\title{Return dates in EViews format}
\usage{
dateAsEV(x)
}
\arguments{
\item{x}{A ts object.}
}
\value{
A character with EViews-compatible dates.
}
\description{
Return dates in EViews format
}
\examples{
x <- ts(1:10, start = c(2001, 1), freq = 4)
y <- ts(1:20, start = c(2001, 1), freq = 12)
dateAsEV(x)
dateAsEV(y)
}
