% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa.R
\name{diagnoseRevisions}
\alias{diagnoseRevisions}
\title{Compute discrepancy / revision summary statistics}
\usage{
diagnoseRevisions(old, new, nspans = NA, overlap.length = NA)
}
\arguments{
\item{old}{Numeric: the series before revision}

\item{new}{Numeric: the series after revision}

\item{nspans}{Numeric: number of spans. Returned in the output, does not affect the result.}

\item{overlap.length}{Numeric: length of one overlap of two spans. Returned in the output, does not affect the result.}
}
\value{
A numeric vector with 5 elements: mean revision, mean absolute revision, mean absolute percentage revision
(known as Mean Avg. % Error in forecasting), SD of revision, relative mean absolute revision (ratio between mean
absolute revision and mean absolute old series), ratio of standard deviations (SD(revision) / SD(old))
}
\description{
Compute discrepancy / revision summary statistics
}
\examples{
before <- c(1.2, 1.2, 0.5, 1.1, 1.7, 0.6, 0.9, 1, 1.7, 0.8)
after <- c(1.3, 1.2, 0.6, 1.3, 2.1, 0.2, 0.8, 1.1, 1.8, 0.8)
diagnoseRevisions(before, after)
}
