% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecm.R
\name{ECM2ARDL}
\alias{ECM2ARDL}
\title{Convert ECM coefficients to an equivalent ARDL}
\usage{
ECM2ARDL(
  dx.coef.list = NULL,
  dy.coef.vec = NULL,
  ECT,
  long.run,
  varnames = NULL
)
}
\arguments{
\item{dx.coef.list}{A list of numeric vectors of coefficients on contemporaneous and lagged regressor differences (\code{Dx[t-0]}, \code{Dx[t-1]}, ...).}

\item{dy.coef.vec}{A numeric vector of coefficients on lagged dep. var. differences (\code{Dy[t-1]}, \code{Dy[t-2]}, ...).}

\item{ECT}{A numeric scalar: coefficient on the ECT (adjustment strength). For stable models, must be between -2 and 0 (0 means that there is no error correction).}

\item{long.run}{A numeric vector of coefficients on the lagged regressor levels (\code{x[t-1]}).}

\item{varnames}{A character vector for human-readable variable names.}
}
\value{
A list of AR coefficient vector and DL coefficient list.
}
\description{
Convert ECM coefficients to an equivalent ARDL
}
\examples{
# Imagine an error-correction model:
# dY = 0.2*dY(-1) + 0.9*dX1 + 0.2*dX2 - 0.8(Y(-1) - 1.5*X1(-1) + 0.5*X2(-1)) + U
# Its ARDL equivalent is
# Y = 0.4*Y(-1) - 0.2*Y(-2) + 0.9*X1 + 0.2*X2 + 0.3*X1(-1) - 0.6*X2(-1) + U
rho <- 0.2
dx <- c(x1 = 0.9, x2 = 0.2)
ECT <- -0.8
lx <- c(x1 = -1.5, x2 = 0.5)
ARDL.coef <- ECM2ARDL(dx.coef.list = dx, dy.coef.vec = rho, ECT = ECT, long.run = lx)
}
