% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{findMiddleGaps}
\alias{findMiddleGaps}
\title{Find non-numeric gaps in the middle}
\usage{
findMiddleGaps(x)
}
\arguments{
\item{x}{Any vector, possibly with NAs, NaNs, Infs etc. in the middle}
}
\value{
Indices of middle gaps. If there are no middle gaps, returns NULL.
}
\description{
Find non-numeric gaps in the middle
}
\examples{
findMiddleGaps(c(rep(NA, 4), 1:10, rep(NA, 3))) # No middle gaps
findMiddleGaps(c(1:10, rep(NA, 3), 5:9)) # One middle gap
findMiddleGaps(c(NA, 2:4, NA, NA, 7:8, NA, 10:12, NA, NA)) # Two middle gaps
}
