% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{findVecPos}
\alias{findVecPos}
\title{Get indices of a vector of numbers in another vector}
\usage{
findVecPos(what, where)
}
\arguments{
\item{what}{A vector to look up.}

\item{where}{A vector in which to look up.}
}
\value{
A vector of indices (length 0 or above).
}
\description{
Get indices of a vector of numbers in another vector
}
\examples{
findVecPos(1:3, c(7, 4, 1, 2, 3, 5))
}
