% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm.R
\name{genDFM}
\alias{genDFM}
\title{Generate EViews code with reasonable initial values}
\usage{
genDFM(
  signals = c("emp_m_g_s", "u_m_g_s", "pib_r_q_g_s"),
  signals.with.1strev = NULL,
  signals.1strev = NULL,
  sig.names = c("Total employment", "Unemployed", "Lux GDP"),
  sig.w1r.names = NULL,
  mod.name = "emp",
  data = NULL,
  nfac = 2,
  err.lag = 2,
  fac.lag = 2,
  rev.lag = FALSE,
  positive.first = TRUE,
  trim.ar = 0.9
)
}
\arguments{
\item{signals}{Character vector of signal names (without first revisions)}

\item{signals.with.1strev}{Character vector of signal names that have a first revision (FR)}

\item{signals.1strev}{Character vector of vector names that are first revisions}

\item{sig.names}{Signal names (without FR) for human-readable output}

\item{sig.w1r.names}{Signal names with first revision for human-readable output}

\item{mod.name}{State-space model name for the code}

\item{data}{Data set to compute the initial values; must contain the column names identical to signal names (signals, signals.with.1strev, signals.1strev).}

\item{nfac}{Number of unobserved factors}

\item{err.lag}{Maximum AR lag of the dynamic observation error (a non-negative integer scalar)}

\item{fac.lag}{Maximum AR lag of the dynamic factor (a non-negative integer scalar)}

\item{rev.lag}{If TRUE, adds one lag of revision error}

\item{positive.first}{If TRUE, enforce most positive coefficients on the first principal component}

\item{trim.ar}{A small positive integer less than 1: if the preliminary estimates of AR coefficients are greater than this number, trim them at this level to ensure stationarity}
}
\value{
A list with 5 elements: EViews model text, tab-separated table for copying into Excel, initial-value vector, the imputed data set, and PCA done on it
}
\description{
Generate EViews code with reasonable initial values
}
\examples{
# A simple model with 3 variables and 2 factors
genDFM()
}
