% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{genRamp}
\alias{genRamp}
\title{Generate a linear ramp from 0 to 1 or possible reversed}
\usage{
genRamp(x, start = 2001.5, end = 2002.5, rev = FALSE)
}
\arguments{
\item{x}{Time series that has the time attribute}

\item{start}{Start time (or if there is no time(x) attribute, index)}

\item{end}{End time (or if there is no time(x) attribute, index)}

\item{rev}{If TRUE, starts at 1 and goes down to zero; otherwise, starts at 0}
}
\value{
A series of the same length of x where the change from 0 to 1 occurs gradually
}
\description{
Generate a linear ramp from 0 to 1 or possible reversed
}
\examples{
genRamp(1:100, 40, 50, rev = TRUE)
}
