% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecm.R
\name{getECT}
\alias{getECT}
\title{Extract the error-correction term from an ECM}
\usage{
getECT(x, yname, xnames, centre = TRUE, detrend = FALSE)
}
\arguments{
\item{x}{An ECM (lm or restriktor).}

\item{yname}{Lagged dependent variable name (e.g. `"lag1_y"`).}

\item{xnames}{Lagged dependent variable names (e.g. `c("lag1_x1", "lag1_x2")`).}

\item{centre}{If TRUE, subtract the mean. Use for unrestricted constant.}

\item{detrend}{If TRUE, subtract the linear trend Use for unrestricted trend}
}
\value{
A numeric vector of long-run residuals (Y[t-1] - X[t-1]'b).
}
\description{
Extract the error-correction term from an ECM
}
