% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{getRange}
\alias{getRange}
\title{Extract the valid range of observations}
\usage{
getRange(data, date.format = c("eviews", "X13", "Date", "time", "index"))
}
\arguments{
\item{data}{A ts or mts object}

\item{date.format}{Output format compatibility. EViews: 2021Q4, X13: 2021.4,
Date: 2021-04-01, time: 2021.25, index: integer position}
}
\value{
A vector or a data frame with the results of getRangeVec() combined
}
\description{
Extract the valid range of observations
}
\examples{
getRange(ts(cbind(x1 = c(NA, NA, 1:3, NA, 1:3, NA, NA, 2, NA), x2 = 1:13),
         start = c(2020, 1), freq = 12), date.format = "Date")
}
