% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa.R
\name{getSAStat}
\alias{getSAStat}
\alias{getSAStatOne}
\title{Get the key seasonal adjustment statistics}
\usage{
getSAStat(x, skip.robustM7 = FALSE, sep = " ")

getSAStatOne(x, skip.robustM7 = FALSE, sep = " ")
}
\arguments{
\item{x}{An object of class [seasonal::seas], or a list of outputs from
[seasonal::seas] or [diagnoseSeasonality()].}

\item{skip.robustM7}{If TRUE, does not test the significance of seasonal and
year dummies with a consistent VCOV matrix. Saves 0.1 seconds in most cases}

\item{sep}{A character used as a delimiter separating multiple outlier entries
(is used as the `collapse` argument of `paste0()`)

`getSAStat()` can be applied to a [seasonal::seas] object or to the output of
[diagnoseSeasonality()] regardless of the number of models / classes.
If an adjustment is blended from multiple models, `getSAStat()` also computes the revision
stability in the overlapping portions.

`getSAStatOne()` is a low-level function for one single [seasonal::seas] object.}
}
\value{
A data frame with named values that are the most important for diagnostics
and archiving.
}
\description{
Get the key seasonal adjustment statistics
}
\examples{
xs1 <- seasonal::seas(AirPassengers, x11 = "")
getSAStat(xs1)
getSAStatOne(xs1)
xs2 <- diagnoseSeasonality(AirPassengers, name = "AirlinePass", verbose = 1)
getSAStat(xs2) # This one has a name
# A list of multiple models, possibly mixed, coming from seas or diagnoseSeasonality
xs3 <- diagnoseSeasonality(AirPassengers, td = 1, leap.year = FALSE, verbose = 1)
set.seed(1)
xs4 <- diagnoseSeasonality(ts(rnorm(120), start = c(1970, 1), freq = 12),
         td = 1, max.length = 7, name = "WhiteNoise")
getSAStat(list(xs1, xs2, xs3, xs4))
}
