% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa.R
\name{getStat}
\alias{getStat}
\title{Get the crucial seasonal adjustment statistics}
\usage{
getStat(x)
}
\arguments{
\item{x}{An object of class seasonal::seas.}
}
\value{
A list with named values.
}
\description{
Get the crucial seasonal adjustment statistics
}
\examples{
xs <- seasonal::seas(datasets::AirPassengers, x11 = "")
getStat(xs)
}
