% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecm.R
\name{guessECM}
\alias{guessECM}
\title{Extract variable names automatically from an ECM}
\usage{
guessECM(mod, d.prefix = "^d_", lag.prefix = "^lag1?_")
}
\arguments{
\item{mod}{An lm object}

\item{d.prefix}{Character: a regex for identifying differenced variables.}

\item{lag.prefix}{Character: a regex for identifying lagged variables.}
}
\value{
A named list with separated ECM components (named vectors).
}
\description{
Extract variable names automatically from an ECM
}
