% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa.R
\name{imputeTS5}
\alias{imputeTS5}
\alias{imputeMTS5}
\alias{imputePanel7}
\title{Reconstruct missing values in time series via multiple methods}
\usage{
imputeTS5(
  x,
  robust.weights = NULL,
  calendar = NULL,
  sample.begin = NULL,
  sample.end = NULL,
  name = "",
  force.positive = FALSE,
  return.multi = FALSE,
  attr.seas = FALSE
)

imputeMTS5(
  x,
  robust.weights = NULL,
  calendar = NULL,
  sample.begin = NULL,
  sample.end = NULL,
  force.positive = FALSE,
  parallel = FALSE,
  cores = 4,
  return.multi = FALSE
)

imputePanel7(
  x,
  trim = c(0.25, 0.25),
  robust.weights = NULL,
  calendar = NULL,
  sample.begin = NULL,
  sample.end = NULL,
  force.positive = FALSE,
  parallel = FALSE,
  cores = 4,
  seed = 1,
  return.multi = FALSE
)
}
\arguments{
\item{x}{A ts object (not mts)}

\item{robust.weights}{# Weights assigned to the 5 or 7 methods.}

\item{calendar}{Passed to X13.}

\item{sample.begin}{Index (e.g. `c(2010, 1)`) to shrink the estimation sample if needed.}

\item{sample.end}{Index (e.g. `c(2025, 4)`) to shrink the estimation sample if needed.}

\item{name}{Passed to diagnoseSeasonality() (series name).}

\item{force.positive}{# If TRUE, does not return negative values; replacing them with half of the minimal positive value.}

\item{return.multi}{If TRUE, return the 5 series (7 for imputePanel7) without any processing}

\item{attr.seas}{If TRUE, attach the best X13 seasonal model as an attribute (useful for diagnostics).}

\item{parallel}{If TRUE, process the series in parallel via parLApplyLB.}

\item{cores}{Integer: the number of cores for a multi-core cluster.}

\item{trim}{Trim these fractions of panel for Amelia back- and forecasting (to have a model more relevant for the sample ends)}

\item{seed}{Seed for Amelia imputation.

Methods used in `imputeTS5`:
* X13-ARIMA-SEATS forecast and backcast
* Kalman filter for a basic structural model (`type = "BSM"` in [stats::StructTS])
* Kalman filter for [forecast::auto.arima]
* [imputeTS::na_seadec]
* [imputeTS::na_seasplit]

`imputePanel7` calls `imputeMTS5` on multi-variate time series and two more
panel-based methods:
* [Amelia::amelia]
* [missMDA::imputePCA]}
}
\value{
imputeTS5: if \code{return.multi}, a 5-column matrix, otherwise an imputed time series.
imputeMTS5/imputePanel7: if \code{return.multi}, a list of 5- or 7-column matrices, otherwise
an imputed MTS.
}
\description{
Reconstruct missing values in time series via multiple methods
}
\examples{
x <- seq(0, 7.99, 1/12)
set.seed(1) # Generating 3 series wit similar dynamics
y1 <- ts(10 + 0.05*x + 5*(x >= 6) - 4*(x == 5) + 40/(x+20)*cos(x*pi*2) +
         arima.sim(n = 96, list(ar = 0.8)), start = c(2010, 1), freq = 12)
y2 <- ts(3 + 0.05*x + 5*(x >= 6) - 4*(x == 5) + 40/(x+20)*sin(x*pi*2) +
         arima.sim(n = 96, list(ar = 0.8)), start = c(2010, 1), freq = 12)
y3 <- ts(3 - 0.05*x + 5*(x < 6) - 4*(x == 5) + 40/(x+20)*cos(x*pi*2) +
         arima.sim(n = 96, list(ar = 0.8)), start = c(2010, 1), freq = 12)
y2[1:11] <- NA # Creating missing observations to impute
y3[87:96] <- NA
y <- ts.union(y1, y2, y3)
yr <- range(y, na.rm = TRUE)
plot(y, plot.type = "single", ylim = yr, col = 1:3)
y2i <- imputeTS5(y2, return.multi = TRUE) # Takes 10--20 seconds
plot(y2i, plot.type = "single", col = 5:1)
legend("top", colnames(y2i), ncol = 5, col = 5:1, lwd = 1)
\donttest{
# This takes 20 seconds per series, i.e. 2 minutes per this section
yi <- imputeMTS5(y)
plot(yi, plot.type = "single", ylim = yr, ylab = "y", col = 1:3)
yp <- imputePanel7(y)
par(new = TRUE)
plot(yp, plot.type = "single", ylab = "", ylim = yr, col = 1:3, lty = 2)
}
}
