% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{labelsWithHalo}
\alias{labelsWithHalo}
\title{Add labels with a white halo for readability}
\usage{
labelsWithHalo(
  x,
  y,
  labels,
  nhalo = 16,
  halo.col = "#FFFFFF66",
  hscale = 0.02,
  vscale = 0.02,
  abs.h = FALSE,
  abs.v = FALSE,
  ...
)
}
\arguments{
\item{x}{Passed to text().}

\item{y}{Passed to text().}

\item{labels}{Passed to text().}

\item{nhalo}{The number of semi-transparent copies of the object to add as a background.}

\item{halo.col}{Colour of halo elements (default: white semi-transparent).}

\item{hscale}{Horizontal radius of the halo ellipse (multiple of current plot width)}

\item{vscale}{Vertical radius of the halo ellipse (multiple of current plot height)}

\item{abs.h}{Logical: use absolute units of x instead of the current width}

\item{abs.v}{Logical: use absolute units of y instead of the current height}

\item{...}{Passed to text(); both the main text and the halos are invoked with it.}
}
\value{
Nothing.
}
\description{
Add labels with a white halo for readability
}
\examples{
set.seed(1)
plot(runif(2000), pch = 16)
points(500, 0.6, pch = 16, col = 2)
labelsWithHalo(500, 0.6, "Point of interest", nhalo = 32, pos = 3)
}
