% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{makeTS}
\alias{makeTS}
\title{Creates periodic time series from an imported data file}
\usage{
makeTS(data, time.regex = "^Date|^Time|^Code")
}
\arguments{
\item{data}{The data frame, or time series, or tibble to convert}

\item{time.regex}{A string representing a regular expression to match the time-identifying variable

This function tries to guess the periodicity (monthly, quarterly) from the
date column, as a human would do, and create regular ts series.}
}
\value{
A ts or mts object.
}
\description{
Creates periodic time series from an imported data file
}
\examples{
set.seed(1)
d <- data.frame(Time = seq.Date(as.Date("2001-01-01"), as.Date("2020-01-01"), "month"),
           X = rnorm(229), Y = rchisq(229, df = 2), Z = runif(229))
dts <- makeTS(d)
plot(dts)
}
