% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{meanSD}
\alias{meanSD}
\title{Compute mean and SD without NAs}
\usage{
meanSD(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A vector of length 2.
}
\description{
Compute mean and SD without NAs
}
\examples{
meanSD(1:10)
}
