% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{myDiff}
\alias{myDiff}
\title{Compute differenced series padded with NA}
\usage{
myDiff(x, lag = 1L, order = 1L)
}
\arguments{
\item{x}{A numeric vector, matrix or data frame}

\item{lag}{An integer: lag order for differencing (e.g. 12 for seasonal differences of monthly data)}

\item{order}{An integer: differencing order}
}
\value{
An object of the same dimensions as the input.
}
\description{
Compute differenced series padded with NA
}
\examples{
myDiff(mtcars)
myDiff(mtcars, lag = 2)
myDiff(mtcars, lag = 3, order = 2)
}
