% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{myLag}
\alias{myLag}
\title{Compute lagged series padded with NA}
\usage{
myLag(x, lag = 1L)
}
\arguments{
\item{x}{A numeric vector, matrix or data frame}

\item{lag}{An integer: lag order}
}
\value{
An object of the same dimensions as the input.
}
\description{
Compute lagged series padded with NA
}
\examples{
cbind(1:10, myLag(1:10))
}
