% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributions.R
\name{negBarPlot}
\alias{negBarPlot}
\title{Plot the positive and negative contributions above and below zero}
\usage{
negBarPlot(
  x,
  xlim = NULL,
  ylim = NULL,
  main = "Contribuiton breakdown",
  width = 1,
  space = 0.2,
  panel.first = NULL,
  warn = TRUE,
  ...
)
}
\arguments{
\item{x}{A matrix that will be passed to \code{barplot}.}

\item{xlim}{Passed to `barplot()`.}

\item{ylim}{Passed to `barplot()`.}

\item{main}{Plot title (used separately to prevent title duplication).}

\item{width}{Passed to `barplot()`.}

\item{space}{Passed to `barplot()`.}

\item{panel.first}{Passed to `plot()`.}

\item{warn}{If `TRUE`, warns the user in case they are about to save graphical output to a device.}

\item{...}{Passed to `barplot()`.}
}
\value{
Invisibly returns a list of two bar plots (for positive and negative values).
}
\description{
Plot the positive and negative contributions above and below zero
}
\examples{
set.seed(1)
negBarPlot(matrix(rnorm(24), ncol = 6), col = rainbow(4, end = 0.7))
}
