% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{orderedWMean}
\alias{orderedWMean}
\title{Compute a weighted mean of ordered matrix rows with friendly weight names}
\usage{
orderedWMean(x, w = NULL, order = NULL)
}
\arguments{
\item{x}{A numeric matrix for which the ordered rows should be averaged with weights}

\item{w}{A numeric vector of weights or name of the weighting function: "mean", "median", "IQM", "Bartlett", "Parzen", "Tukey-Hanning", "trimean", "midhinge"}

\item{order}{A numeric vector of column orders or numeric matrix of orders within each row.

By default, uses triangular (Bartlett) weights.
If order = NULL, uses apply(x, 1, order) values to order the values in rows of x.
If order is a vector, uses these to sort the columns of x.
If order is a matrix, uses the rows of order to sort the rows of x.}
}
\value{
A numeric vector of weighted mean of ordered x.
}
\description{
Compute a weighted mean of ordered matrix rows with friendly weight names
}
\examples{
set.seed(1)
x <- matrix(rt(20*9, df = 1), ncol = 9) # series with many outliers
matplot(x, type = "l", lty = 1)
xm <- cbind(orderedWMean(x, w = c("mean", "iqm", "median", "trimean", "Bartlett")),
            omit.ends = orderedWMean(x, w = c(0, rep(1, ncol(x)-2), 0)))
mycol <- rainbow(6, v = 0.7, end = 0.8, alpha = 0.8)
matplot(xm, type = "l", col = mycol, lty = 1, lwd = 2)
legend("topleft", colnames(xm), ncol = 2, lwd = 2, col = mycol)
}
