% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecm.R
\name{plotContribECM}
\alias{plotContribECM}
\title{Visualise the past and present contributions in error-correction models}
\usage{
plotContribECM(
  x,
  single.plot = TRUE,
  col = NULL,
  main = "Contribution decomposition",
  dep.var.name = "dep. var.",
  las = NULL,
  left.bar.skip = NULL,
  leg.pos = "topleft",
  mar = NULL,
  ...
)
}
\arguments{
\item{x}{A list returned by `printContrib(, type = "twocolumn")` with two matrices with identical dimensions.
The first matrix is assumed to contain short-run contributions, the second one -- long-run ones.}

\item{single.plot}{If `TRUE`, produces a single bar plot instead of three vertically stacked ones.}

\item{col}{A vector of colours for the bar elements.}

\item{main}{Character: if `single.plot` is `TRUE`, the plot title.}

\item{dep.var.name}{Character: the variable name to be added to the title.}

\item{las}{Integer: label orientation (passed to `axis()`). If `NULL`, uses `1` (normal) for single plots and `2` (90 CCW) for multiple plots.}

\item{left.bar.skip}{Numeric: how many full column widths to skip? Defaults to extra 25\% space.}

\item{leg.pos}{Character: legend position (passed to `legend()`),}

\item{mar}{Numeric of length 4: passed to `par()` (margins for plots)}

\item{...}{Passed to `negBarPlot()`.

The multiple plots are always plotted in the same vertical scale (for comparability).}
}
\value{
Nothing (invisible NULL).
}
\description{
Visualise the past and present contributions in error-correction models
}
\examples{
# This function works with arbitrary matrices
set.seed(1)
xn <- c("GDP", "Empl", "Infl")
x <- list(matrix(rnorm(12), 3, 4, dimnames = list(xn, 2019:2022)),
          matrix(rnorm(12), 3, 4, dimnames = list(xn, 2019:2022)))
plotContribECM(x)
}
