% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributions.R
\name{plotContribLM}
\alias{plotContribLM}
\title{Visualise contributions from many linear models}
\usage{
plotContribLM(
  x,
  name.replacement.mat = NULL,
  contrib.threshold = 0.03,
  threshold.type = c("relative", "absolute"),
  small.contrib = c("drop", "mergekeep", "mergedrop", "preserve"),
  small.name = "Autres",
  francais = FALSE,
  plot.diff = TRUE,
  col = NULL,
  left.bar.skip = 3,
  mar = c(5, 2.2, 2.2, 0.2),
  mtext.line = -1,
  ...
)
}
\arguments{
\item{x}{A list returned by computeContribManyLM}

\item{name.replacement.mat}{Character matrix with two columns containing the existing and the new names in its columns}

\item{contrib.threshold}{Numeric: Do not plot the variables whose max. relative contribution to the levels was less than this}

\item{threshold.type}{Character: Drop small contributions based on minimal fraction or absolute threshold?}

\item{small.contrib}{Character: What to do with the contributions below the threshold: ignore, add up, or add up and check if they exceed the threshold, or do nothing}

\item{small.name}{Character: label for all aggregated insignificant contributions}

\item{francais}{If TRUE, the legends are in French}

\item{plot.diff}{If TRUE, adds a bottom plot with the changes}

\item{col}{Character vector Colours of the vars}

\item{left.bar.skip}{How many bars of space to skip on the left (to help the legend fit)}

\item{mar}{Plot margins, par("mar"); change if the x-axis labels won't fit}

\item{mtext.line}{A line under the level plot at which the note is placed.}

\item{...}{Passed to negBarPlot.}
}
\value{
Nothing.
}
\description{
Visualise contributions from many linear models
}
\examples{
set.seed(1)
x1 <- rnorm(100) + 1:100/50 + 1
x2 <- rnorm(100) - 1:100/50 - 1
x3 <- rnorm(100)
y <- 1 + x1 + x2 + 0.5*x3 + rnorm(100)
d <- data.frame(y, x1, x2, x3)
mod1 <- lm(y ~ x1 + x2, data = d)
mod2 <- lm(y ~ x1 + x3, data = d)
d2 <- d[71:100, ]
ctb.one <- computeContribLM(mod1, newdata = d2)
ctb.many <- computeContribManyLM(list(mod1, mod2), newdata = d2)
plotContribLM(ctb.one)
plotContribLM(ctb.many, mtext.line = 2.5)
}
