% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa.R
\name{plotSeas}
\alias{plotSeas}
\title{An improved seasonal plot with extra useful information}
\usage{
plotSeas(
  x,
  name = NULL,
  sa.custom = NULL,
  extend.xlim = c(0, 0),
  skip.boxplot = FALSE,
  skip.robustM7 = FALSE,
  span.index = NULL
)
}
\arguments{
\item{x}{An object of class "seas" returned by [seasonal::seas] or the output
of [diagnoseSeasonality()]}

\item{name}{A name to be put on the plot}

\item{sa.custom}{A `ts` of the adjusted series if they differ from the default
X13 ones. Useful if the user obtains their SA series from elsewhere or decides
that no SA adjustment (only pre-adjustment) should be done.}

\item{extend.xlim}{The number of years to add at the beginning and the end to extend
the plotting range. Useful if the plot should be produced for a larger sample
than the effective sample on which the adjustment was done.}

\item{skip.boxplot}{If `TRUE`, does not produce the box-and-whisker plot lines}

\item{skip.robustM7}{If `TRUE`, does not produce the HAC-robust M7}

\item{span.index}{The index of the plot in a combined adjustment to show.
If NULL, plots the combined series in one plot. If "last", or 0, or -1, plots the last span.}
}
\value{
Nothing (NULL).
}
\description{
An improved seasonal plot with extra useful information
}
\examples{
x <- diagnoseSeasonality(AirPassengers, transform = FALSE, td = 0,
        easter = TRUE, leap.year = TRUE, plot = NA)
x.force <- diagnoseSeasonality(AirPassengers, transform = FALSE, td = 0,
             easter = TRUE, leap.year = TRUE, plot = NA, force.annual = "regress")
x.alt <- seasonal::seas(AirPassengers, x11 = "")
plotSeas(x)
plotSeas(x.force)
plotSeas(x, sa.custom = x.alt$series$d11, skip.boxplot = TRUE)
# Multiple spans
xm <- diagnoseSeasonality(AirPassengers, max.length = 5, transform = "yes")
plotSeas(xm)
}
