% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{plotTSAnimationFrame}
\alias{plotTSAnimationFrame}
\title{Draw 2D time series with evolution made obvious}
\usage{
plotTSAnimationFrame(
  x,
  i = NULL,
  thicker = 48,
  vgrid = NULL,
  hgrid = NULL,
  ...
)
}
\arguments{
\item{x}{A two-dimensional TS object}

\item{i}{An index at which row to stop. In NULL, plots the full range}

\item{thicker}{The number of points at the end to make thicker}

\item{vgrid}{Vertical grid points}

\item{hgrid}{Horizontal grid points}

\item{...}{Passed to plot() (axis labels etc.) before the line drawing is invoked

This function can be run in a loop over \code{i}, and the resulting frames can be saved as an animation.}
}
\value{
Nothing.
}
\description{
Draw 2D time series with evolution made obvious
}
\examples{
# Visualising a co-integrating relationship
set.seed(1)
x <- cumsum(rnorm(108))
y <- x + rnorm(108)
d <- ts(cbind(x, y), start = c(1995, 1), frequency = 4)
plotTSAnimationFrame(d)
}
