% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecm.R
\name{plotWithForecast}
\alias{plotWithForecast}
\title{Plot observed and forecast series with a distinction}
\usage{
plotWithForecast(
  vars,
  vars.fc,
  data,
  expand.ylim = 0.1,
  x = NULL,
  col = NULL,
  centre = TRUE,
  scale = TRUE,
  ...
)
}
\arguments{
\item{vars}{A character vector of variable names to plot as solid lines.}

\item{vars.fc}{A character vector of forecast variables names to plot as dashed lines.}

\item{data}{A aata frame containing all the \code{vars} and \code{vars.fc} variables.}

\item{expand.ylim}{A numeric scalar: increase the bottom white space for the legend by this share of ylim}

\item{x}{A numeric vector of time-dimension values (usually the time). If `NULL`, integer indices are used.}

\item{col}{A character vector of line colours}

\item{centre}{If TRUE, subtract the mean of each variable.}

\item{scale}{If TRUE, divide by the SD of each variable.}

\item{...}{Passed to plot.}
}
\value{
Nothing (invisible NULL).
}
\description{
Plot observed and forecast series with a distinction
}
\examples{
set.seed(1)
xfc <- matrix(cumsum(rnorm(120)), ncol = 4) # Complete data
x <- xfc
x[c(26:30, 57:60, 89:90, 111:120)] <- NA # Adding incompleteness
colnames(x) <- paste0("X", 1:4); #' colnames(xfc) <- paste0("fX", 1:4)
d <- cbind(x, xfc)
plotWithForecast(vars = colnames(x), vars.fc = colnames(xfc), data = d)
}
