% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecm.R
\name{prepareTransform}
\alias{prepareTransform}
\title{Compute lags and differences of a variable for an ECM}
\usage{
prepareTransform(varname, data, nlag = 2)
}
\arguments{
\item{varname}{Character: variable name}

\item{data}{Data frame where this variable can be found}

\item{nlag}{Number of lags to compute}
}
\value{
A data frame with the lags and differences
}
\description{
Compute lags and differences of a variable for an ECM
}
\examples{
d <- cbind(X1 = rnorm(100), X2 = cumsum(rnorm(100)))
prepareTransform("X2", d)
}
