% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm.R
\name{printDFM}
\alias{printDFM}
\title{Print the EViews code of a DFM and a tab-separated label list}
\usage{
printDFM(x)
}
\arguments{
\item{x}{List returned by genDFM}
}
\value{
Nothing.
}
\description{
Print the EViews code of a DFM and a tab-separated label list
}
\examples{
m <- genDFM()
printDFM(m)
}
