% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecm.R
\name{getLR}
\alias{getLR}
\alias{printECM}
\alias{getECM}
\title{Print one or multiple ECMs in an easily interpretable format}
\usage{
getLR(x, yname, xnames)

printECM(
  ...,
  yname,
  xnames,
  d.prefix = "d_",
  lag.prefix = "lag1_",
  mod.names = NULL,
  shorten = function(x) formatC(x, format = "f", digits = 2, flag = " "),
  select = c("ec", "long", "short", "inertia", "intercept"),
  minus = TRUE
)

getECM(
  x,
  yname,
  xnames,
  d.prefix,
  lag.prefix,
  na.to.zero = FALSE,
  incorporate.intercept = FALSE
)
}
\arguments{
\item{x}{An lm model or a numeric vector}

\item{yname}{A string with the dependent variable name.}

\item{xnames}{A character vector of explanatory variable names that appear in both parts and differ by prefixed. Unique variables (like time dummies) need not be included.}

\item{...}{Estimated model outputs from lm, restriktor::restriktor or restriktor::conLM.}

\item{d.prefix}{A string that should precede every differenced variables (usually d_ or U+0394 if the locale is Unicode)}

\item{lag.prefix}{A string that should to precede first lags.}

\item{mod.names}{A character vector of model names for printing.}

\item{shorten}{A function to format numbers (usually round, sprintf or formatC).}

\item{select}{A character vector showing which parts of the equation to keep for reporting.}

\item{minus}{Logical: multiply the OLS long-run estimates by -1?}

\item{na.to.zero}{If TRUE, replace NAs with zeros.}

\item{incorporate.intercept}{If TRUE, report the intercept in the short-run part.}
}
\value{
Invisibly returns the matrices that have just been shown (without rounding).
}
\description{
Print one or multiple ECMs in an easily interpretable format
}
\section{Functions}{
\itemize{
\item \code{getLR()}: Extract the long-run elasticities

\item \code{getECM()}: Internal function to sort ECM components

}}
