% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecm.R
\name{printShort}
\alias{printShort}
\title{Prints hypothesis testing results in a concise manner}
\usage{
printShort(x, hyp = NULL, digits = 3)
}
\arguments{
\item{x}{An object of class anova (from car::linearHypothesis) or htest (from nlWaldTest::nlWaldlest)}

\item{hyp}{Optional: human-readable hypothesis name}

\item{digits}{Number of p-values digits to print.}
}
\value{
Nothing (invisible NULL).
}
\description{
Prints hypothesis testing results in a concise manner
}
