% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{printStars}
\alias{printStars}
\title{Print stars based on the coefficients and standard errors}
\usage{
printStars(coef, se, lev = c(0.1, 0.05, 0.01), add.number = TRUE, digits = 3)
}
\arguments{
\item{coef}{A numeric vector representing coefficients}

\item{se}{A numeric vector representing standard errors}

\item{lev}{A numeric vector of p-values that determine the number of stars}

\item{add.number}{Logical print number with the p-values? If FALSE, returns just the stars}

\item{digits}{How many digits of the number to print}
}
\value{
A character vector of stars or numbers with stars
}
\description{
Print stars based on the coefficients and standard errors
}
\examples{
printStars(coef = seq(1, 3, 0.2), se = rep(1, 11), digits = 1)
}
