% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{printSym}
\alias{printSym}
\title{Print a line of N characters}
\usage{
printSym(x = "=", times = 60)
}
\arguments{
\item{x}{A character to print.}

\item{times}{How many times to print it.}
}
\value{
Nothing.
}
\description{
Print a line of N characters
}
\examples{
printSym("*", 30)
}
