% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{printTab}
\alias{printTab}
\title{Print a table to Excel or TeX}
\usage{
printTab(
  x,
  format = c("excel", "TeX"),
  digits = 3,
  zero.replace = TRUE,
  minus.replace = TRUE,
  nozero.integers = TRUE,
  nozero.tol = 1e-06,
  include.rownames = FALSE,
  na = ""
)
}
\arguments{
\item{x}{A data frame or a matrix to print}

\item{format}{Character: if "excel", optimises the plain-text look; if TeX, produces a LaTeX table}

\item{digits}{Integer: Number of decimal positions to print}

\item{zero.replace}{Logical: replace true zeros with a dash?}

\item{minus.replace}{Logical: replace the minus dash with a proper typographic minus sign?}

\item{nozero.integers}{Logical: if TRUE, do not add zeros to the numbers very close to integers}

\item{nozero.tol}{A number is considered integer and not padded with zeros if it is closer to the nearest integer than this tolerance}

\item{include.rownames}{Logical: add the row names as the first column?}

\item{na}{String to use instead of NAs.}
}
\value{
Nothing.
}
\description{
Print a table to Excel or TeX
}
\examples{
# This output is ready to be copied from the console and paste into Excel
printTab(mtcars, digits = 2, zero.replace = FALSE, minus.replace = FALSE)
printTab(myDiff(datasets::airquality[1:10, ]), digits = 1,
         zero.replace = FALSE, format = "TeX", na = "X")
}
