% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{quiet}
\alias{quiet}
\title{Run a command quietly}
\usage{
quiet(x)
}
\arguments{
\item{x}{Any R expression.

https://stackoverflow.com/a/54136863}
}
\value{
The same result as evaluated by x, only with no output (sinked into tempfile()).
}
\description{
Run a command quietly
}
\examples{
quiet(print("Test"))
quiet(cat("Test\n"))
quiet(message("Test")) # Still prints a message
}
