% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{readAllSheets}
\alias{readAllSheets}
\title{Load all Excel sheets as a list of vectors}
\usage{
readAllSheets(
  xlsxFile,
  verbose = TRUE,
  vector = TRUE,
  lowercase = TRUE,
  intercept2c = TRUE,
  ...
)
}
\arguments{
\item{xlsxFile}{Character: path to the Excel file.}

\item{verbose}{If TRUE, prints the loading progress (useful if the file is on the network drive).}

\item{vector}{If TRUE, returns a list of vectors; otherwise, a list of data frames.}

\item{lowercase}{If TRUE, converts the first column (supposedly containing the names) to lowercase.}

\item{intercept2c}{If TRUE, replaces all instances of `"c"` with `"(Intercept)"` in the first column.}

\item{...}{Passed to `openxlsx::read.xlsx` that is applied to each detected sheet.

To simply load the XLSX file as a list of sheets, invoke the function with
`vector`, `lowercase`, and `intercept2c` set to `FALSE`.}
}
\value{
A list of named vectors or (if `vector` is `FALSE`, a list of data frames from the sheets).
}
\description{
Load all Excel sheets as a list of vectors
}
\examples{
\dontrun{
# To compute contributions, load an Excel file where the 1st column
# of each sheet contains names, and the second one, coefficient values
readAllSheets("myfile.xlsx")

# Simply load a file as a list of sheets
readAllSheets("myfile.xlsx", vector = FALSE, lowercase = FALSE, intercept2c = FALSE)
}
}
\seealso{
[openxlsx::read.xlsx()] for everything that can be passed to `...`.
}
