% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa.R
\name{robustSeasTests}
\alias{robustSeasTests}
\title{Robustified identifiable and stable seasonality tests for X13 M7}
\usage{
robustSeasTests(x, type = c("rlm", "lm", "rank", "nonrobust"))
}
\arguments{
\item{x}{An object of class "seas" returned by [seasonal::seas]}

\item{type}{Character: "rlm"for a robust linear model (default), "lm" for a linear model,
"rank" for the rank regression, or "nonrobust" for the homoskedasticity-based ANOVA.

This function provides a long-overdue improvement for the X13 ANOVA-based tests
due to Higgins (1975)
by using (1) weighted robust regression instead of the rank regression and
(2) heteroskedasticity-and-autocorrelation-consistent VCOV instead of the
unreaslistic white-noise assumption that almost never holds for the irregular
component.

The Kruskal--Wallis and Friendman are non-robust to the violation of the
homoskedasticity assumption. Insted, we run two simple linear regressions
and test the joint significance of the dummies of interest.

Define SI as the sum of the seasonal and the irregular components.

1. Identifiable seasonality: regress SI on the seasonal dummies

If some periods exhibit huge seasonal effects and/or atypical values, then,
the rank regression (`type = "rank"`) would ignore their magnitude that could
potentially translate into stronger #' estimated seasonality effect.
If `type = "rlm"`, then, IRLS via [MASS::rlm()] reduces the impact of such
outliers to some extent.
HAC-robust ANOVA is conducted in all cases. HAC estimation may fails if
there are singular year dummies. In this case, such singular observations
are dropped. If still, the model matrix is singular, it reverts to
non-robust VCOV estimation with a warning.

If IRLS does not converge, the fallback is the rank regression with HAC inference.
If HAC cannot be computed, the fallback is the HC0 and, should it fails as well,
the non-robust VCOV (although in practice, it should never be the case).}
}
\value{
A list with 4 components: `models` (the list of 2 models), `tests` (the list of 2 ANOVAs),
`robust.m7` (the final robust M7 statistic), and the HAC bandwidth used in VCOV estimation; if it failed
due to singularities, bandwidth is `NA`.
}
\description{
Robustified identifiable and stable seasonality tests for X13 M7
}
\examples{
x1 <- AirPassengers
x2 <- window(x1, start = c(1949, 12), end = c(1960, 1)) # With singular years
sa1 <- seasonal::seas(x1, x11 = "")
sa2 <- diagnoseSeasonality(x2)
# getStat(list(sa1, sa2))
robustSeasTests(sa1)
robustSeasTests(sa2$seas)
}
