% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecm.R
\name{sliceList}
\alias{sliceList}
\title{Generate a list of sliding-window data frames}
\usage{
sliceList(x, start = NULL, end = NULL)
}
\arguments{
\item{x}{A data frame, matrix, or numeric vector.}

\item{start}{Integer: indices of the sample starts.}

\item{end}{Integer: indices of the sample ends

Create a list of data frames, matrices, or vectors, each trimmed to the
corresponding starts and ends.}
}
\value{
A list of data frames trimmed to the start and end, with index attributes.
}
\description{
Generate a list of sliding-window data frames
}
\examples{
x <- matrix(rnorm(100), ncol = 2)
a1 <- sliceList(x, start = c(1, 11, 21), end = c(30, 40, 50)) # Sliding window
lapply(a1, attr, "indices")
a2 <- sliceList(x, end = c(30, 40, 50)) # Expanding window
lapply(a2, attr, "indices")
}
