% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa.R
\name{splitOverlapFixed}
\alias{splitOverlapFixed}
\alias{splitOverlapCustom}
\title{Compute indices and weights for sub-series with overlapping}
\usage{
splitOverlapFixed(n, nmax, l, plot = FALSE)

splitOverlapCustom(
  n,
  est.starts,
  est.ends,
  blend.starts = NULL,
  blend.ends = NULL,
  plot = FALSE
)
}
\arguments{
\item{n}{Integer: length of the full series}

\item{nmax}{Integer: chunk size}

\item{l}{Non-negative integer: overlap length; can be 0 for no overlap}

\item{plot}{If `TRUE`, visualises the weight matrix and the sub-samples

When a time series is estimated in chunks, the final results should be
merged with ts.union. Then, they can be simply multipled by the weights
and added by row.}

\item{est.starts}{Integer vector of indices denoting the starts of each chunk; should start with 1}

\item{est.ends}{Integer vector of indices denoting the ends of each chunk; should end with `n`
Must have the same length as `est.starts`.}

\item{blend.starts}{Integer vector of indices denoting the starts of blending periods between 2 estimation windows.
Must be of length 1 shorter than est.starts (because `k` chunks = `k-1` cuts between them), must start at 2 or later.
Must not be less that the corresponding element of `est.starts` (excluding the 1st
element): one cannot blend something with nothing; blending must occur within the overlap between the estimations.}

\item{blend.ends}{Integer vector of indices denoting the ends of blending periods between 2 estimation windows.
Must be of length 1 shorter than est.ends, must end at `n-1` or earlier.
Must not be greater that the corresponding element of `est.ends` (excluding the last element).}
}
\value{
A list of two matrices: sub-sample indices by columns (with `NA`s for
unused indices) and weights for blending.
}
\description{
Compute indices and weights for sub-series with overlapping
}
\examples{
splitOverlapFixed(n = 43, nmax = 20, l = 4, plot = TRUE)
# Custom: estimate on 5 and 7 years, blend in between
splitOverlapCustom(n = 96, est.starts = c(1, 13), est.ends = c(60, 96), plot = TRUE)
# Custom: estimate on years 1-5 and 2-7 years, but blend during the year 3
splitOverlapCustom(n = 96, est.starts = c(1, 13), est.ends = c(60, 96),
                   blend.starts = 25, blend.ends = 36, plot = TRUE)
splitOverlapCustom(n = 150, est.starts = c(NA, 45, 90), est.ends = c(50, 100, 150), plot = TRUE)
splitOverlapCustom(n = 200, est.starts = c(1, 15, 90, 140), est.ends = c(50, 100, 139, NA),
                  blend.starts = c(15, 90, NA), blend.ends = c(27, 100, NA), plot = TRUE)
# The ends can be trimmed
splitOverlapCustom(n = 200, est.starts = c(8, 15, 90, 140), est.ends = c(50, 100, 139, 190),
                  blend.starts = c(15, 90, NA), blend.ends = c(27, 100, NA), plot = TRUE)
# If there is anything wrong with the indices, it will throw a meaningflu error
\dontrun{
# Does not work: gap in the middle
splitOverlapCustom(n = 100, est.starts = c(1, 51), est.ends = c(49, 100), plot = TRUE)
# Works: no gap in the middle
splitOverlapCustom(n = 100, est.starts = c(1, 51), est.ends = c(50, 100), plot = TRUE)
# Does not work because the blending window must be within the estimation sample overlap
splitOverlapCustom(n = 150, est.starts = c(NA, 45, 90), est.ends = c(50, 100, NA),
                   blend.starts = c(44, 89), blend.ends = c(51, 101), plot = TRUE)
}
}
