% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa.R
\name{splitWithOverlaps}
\alias{splitWithOverlaps}
\title{Compute indices and weights for sub-series with overlapping}
\usage{
splitWithOverlaps(n, k, l, plot = FALSE)
}
\arguments{
\item{n}{Integer: length of the full series}

\item{k}{Integer: chunk size}

\item{l}{Non-negative integer: overlap length; can be 0 for no overlap}

\item{plot}{If `TRUE`, visualises the weight matrix and the sub-samples

When a time series is estimated in chunks, the final results should be
merged with ts.union. Then, they can be simply multipled by the weights
and added by row.}
}
\value{
A list of two matrices: subsample indices by columns (with `NA`s for
unused indices) and weights for blending.
}
\description{
Compute indices and weights for sub-series with overlapping
}
\examples{
splitWithOverlaps(n = 43, k = 20, l = 4, plot = TRUE)
}
