% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecm.R
\name{testActiveConstraints}
\alias{testActiveConstraints}
\title{Find and test the active constraints in a model}
\usage{
testActiveConstraints(rmod, eps = 1e-04, vcov = NULL, denominator.vars = NULL)
}
\arguments{
\item{rmod}{A restricted model produced by 'restriktor'.}

\item{eps}{Numerical difference epsilon to test the sensitivity of p-values}

\item{vcov}{A variance-covariance matrix for the unrestricted model or a function to apply to the unconstrained model}

\item{denominator.vars}{Optional; for testing ratio hypotheses, the names of variables found in the denominator
Should start with the lag of Y, then the lags of long-run X, e.g. c("lagY", "lagCapital"); whichever is found first
will be put into the denominator, and in the LR testing, the lag of Y is always in the denominator

If no variance-covariance matrix is supplied, uses [sandwich::vcovHAC()]. If
the inference is non-standard (e.g. fast convergence rates in error-correction
models), then, the conclusions might be wrong.}
}
\value{
A numeric matrix with the test p-values and their derivatives
}
\description{
Find and test the active constraints in a model
}
\examples{
m <- lm(mpg ~ factor(cyl) + disp + hp + drat + wt, data = datasets::mtcars)
cns <- c("disp < 0", "hp < 0", "drat > 1", "wt < -6", "wt > -100")
r <- restriktor::restriktor(m, paste0(cns, collapse = "\n"))
# Some of these constraints are inactive; some are tighter than others
testActiveConstraints(r, vcov = sandwich::vcovHC)
}
