% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{time2ind}
\alias{time2ind}
\alias{ind2time}
\alias{date2ind}
\alias{ind2date}
\alias{ts2date}
\title{A set of functions for date conversion}
\usage{
time2ind(x, freq)

ind2time(x, freq)

date2ind(x, freq)

ind2date(x, freq)

ts2date(x)
}
\arguments{
\item{x}{Source vector or matrix of the type described below.}

\item{freq}{Frequency (4 or 12).

time2ind: ts --> (year, cycle): 2012.4167 -> c(2012, 5)
ind2time: (year, cycle) --> ts: c(2012, 5) -> 2012.4167
date2ind: Date --> (year, cycle): "2012-03-21" -> c(2012, 3) or c(2012, 1) (quarterly)
ind2date: (year, cycle) --> Date: c(2012, 3) --> "2012-03-01" or "2012-07-01" (quarterly)
ts2date: ts --> Date

These functions support matrix input (i.e. ind2time accepts 2-column matrices).}
}
\value{
A converted vector representing the time.
}
\description{
A set of functions for date conversion
}
\section{Functions}{
\itemize{
\item \code{ind2time()}: Pair(s) of indices to decimal time

\item \code{date2ind()}: Date to (year, period) index

\item \code{ind2date()}: Time series to (year, period) index

\item \code{ts2date()}: Time series to Date

}}
\examples{
x <- ts(1:24, start = c(2001, 1), freq = 12)
d <- seq(as.Date("2009-01-01"), as.Date("2010-04-01"), by = "week")
i <- cbind(rep(2010:2014, each = 4), rep(1:4, 5))
time2ind(x)
ts2date(x)
date2ind(d, freq = 12)
ind2date(i, freq = 4)
ind2time(i, freq = 4)
}
