% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecm.R
\name{varImportance}
\alias{varImportance}
\alias{plotImportance}
\alias{plotCoefBetter}
\title{Relative importance of regressors in the Elastic Net path}
\usage{
varImportance(x, max.vars = NULL, sorted = TRUE)

plotImportance(vi, col = NULL, ...)

plotCoefBetter(
  glmnet,
  xvar = c("lambda", "norm", "dev"),
  label = TRUE,
  drop.intercept = TRUE,
  hscale = 0.005,
  vscale = 0.01,
  ...
)
}
\arguments{
\item{x}{An estimated glmnet object.}

\item{max.vars}{An integer: how many variables to return. If NULL, returns all model variables.}

\item{sorted}{If TRUE, sorts the variable by relative importance, otherwise use the original name order}

\item{vi}{A vector or a matrix with variable names as rows and percentages in columns.}

\item{col}{A vector of colours (for each columnif `vi` is a matrix).}

\item{...}{Passed to matplot.

Based on glmnet:::plotCoef.}

\item{glmnet}{A glmnet object.}

\item{xvar}{Indicator to plot: penalty factor, norm, or deviance.}

\item{label}{If TRUE, add labels with halo.}

\item{drop.intercept}{If TRUE, do not plot the intercept regularisation path.}

\item{hscale}{Passed to [labelsWithHalo()].}

\item{vscale}{Passed to [labelsWithHalo()].}
}
\value{
A numeric vector with popularity values: 0 means that Elastic Net drops the variable even with small penaltiles, values close to 1 means that this variable was retained in the model for most of the path.

Nothing (invisible NULL).

Nothing.
}
\description{
Relative importance of regressors in the Elastic Net path

Plot glmnet coefficients with better visualisation parameters
}
\examples{
x <- model.matrix(mpg ~ ., data = datasets::mtcars)[, -1]
y <- datasets::mtcars[, "mpg"]
lasso <- glmnet::glmnet(x = x, y = y)
elnet <- glmnet::glmnet(x = x, y = y, alpha = 0.5)
vi1 <- varImportance(lasso); vi2 <- varImportance(elnet)
plotImportance(cbind(vi1, vi2), pch = 1)
plotCoefBetter(lasso)
}
