% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{wrapFormula}
\alias{wrapFormula}
\title{Quickly create a formula y ~ x1 + x2 + ...}
\usage{
wrapFormula(y, x)
}
\arguments{
\item{y}{The left-hand-side variable (character scalar)}

\item{x}{The right-hand-side variables (character vector)}
}
\value{
A formula with the requested variables.
}
\description{
Quickly create a formula y ~ x1 + x2 + ...
}
\examples{
wrapFormula("mpg", c("cyl", "hp", "wt"))
}
