% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa.R
\name{getCalendars}
\alias{getCalendars}
\title{Read national calendars into the global environment.}
\usage{
getCalendars(country = "Luxembourg", suffix = "lu", path = NULL)
}
\arguments{
\item{country}{Character: file name (without the "-Q.csv") prefix}

\item{suffix}{Character: the extra string inserted in the name to distinguish
calendars: the output will be saved as `cal.SUFFIX.m` and `cal.SUFFIX.q`}

\item{path}{Character: full directory path to use (ending with a slash).}
}
\value{
Nothing; loads two MTS into the global environment.
}
\description{
Read national calendars into the global environment.
}
\examples{
tryCatch(getCalendars("France"), error = function(e) print("Calendar not found."))
}
