% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecm.R
\name{simARDL}
\alias{simARDL}
\title{Simulate an ARDL process}
\usage{
simARDL(endog.lag, exog.lag, intercept = 20, seed = 1, n = 300)
}
\arguments{
\item{endog.lag}{Numeric: coefficients on the lags of Y}

\item{exog.lag}{A list of coefficients on X and its lags}

\item{intercept}{Numeric: shift for the entire series}

\item{seed}{Integer: seed for PRNG}

\item{n}{Number of observations}
}
\value{
A data frame with the dependent variable, regressors, and errors
}
\description{
Simulate an ARDL process
}
\examples{
b <- list(AR = c(0.4, -0.2), DL = list(L0 = c(0.9, 0.2), L1 = c(0.3, 0.6)))
d <- simARDL(endog.lag = b$AR, exog.lag = b$DL, n = 40)
d <- ts(d, start = c(2000, 1), freq = 4)
plot(d)
}
